requirements.txt
pytest
fastapi==0.103.2
psycopg2==2.9.1
flask-wtf==0.14.3
psycopg2-binary==2.9.1
werkzeug==0.16.1
markupsafe==2.0.1
Flask-RESTful
Flask-Swagger-UI
flasgger==0.5.4
unittest
requests